auth_n_auto/requirements.txt
Django==5.2.5
djangorestframework==3.16.0
drf-yasg==1.21.10
